реализован обработчик, возвращающий коллекцию вопросов по запросу с параметрами

+++ src/DevQuestions.Contracts/Questions/Responses/QuestionsResponse.cs
@@ -2,4 +2,4 @@ using DevQuestions.Contracts.Questions.Dtos;
 
 namespace DevQuestions.Contracts.Questions.Responses;
 
-public record QuestionsResponse(IEnumerable<QuestionDto> Questions, int TotalCount);
+public record QuestionsResponse(IEnumerable<QuestionDto> Questions, long TotalCount);
